Passes part 2

+++ src/main/scala/aoc2023/day06/BoatRace.scala
@@ -1,12 +1,13 @@
 package aoc2023.day06
 
-case class BoatRace(duration: Int, record: Int) {
-  def winningButtonPresses(): Int = (0 until duration).map(buttonHeld => Boat(duration).distance(buttonHeld)).filter(distance => distance > record).size
+case class BoatRace(duration: Int, record: Long) {
+
+  def winningButtonPresses(): Long = (0 until duration).map(buttonHeld => Boat(duration).distance(buttonHeld)).filter(distance => distance > record).size
 
 }
 
-case class Boat(raceLasts: Int) {
+case class Boat(raceLasts: Long) {
 
-  def distance(buttonHeld: Int): Int = buttonHeld * (raceLasts - buttonHeld)
+  def distance(buttonHeld: Long): Long = buttonHeld * (raceLasts - buttonHeld)
 
 }
